conditionals fixed (I think)
So now there will never be two students with the same email in the database, and there will never be more than one teacher object per user.  The only condition that I'm fuzzy on is what happens when a different account creates a student that is already in the DB.  Ideally, it will append the (already created) student to a list of the teachers students.  I could not test this, need to work on this with a team member