bulk add kinda working!
There is an issue with the code runnign multiple times more than it should.  This is resulting in there being 2-4 iterations of the same userid for each student object, but this causes no harm to the website.  The classPage function that populates the homepage checks to make sure that it only creates one of each student profile max.